added action to remove question
