+++ src/app/(auth)/manager/organizations/[organizationId]/form-templates/[templateId]/questions/_actions.ts
@@ -1,3 +1,6 @@
+"use server";
+
+import { actionFlash } from "@/action-flash";
 import api from "@/api";
 import { ApiResponse } from "@/schemas";
 
@@ -14,6 +17,51 @@ export async function fetchQuestionsAction(
   });
 
   const json = await response.json();
+
+  return ApiResponse.parse(json);
+}
+
+export async function createQuestionAction(
+  templateId: string,
+  prevState: any,
+  formData: FormData,
+): Promise<ApiResponse> {
+  const response = await api({
+    input: `/manager/form-templates/${templateId}/short-questions`,
+    init: {
+      method: "POST",
+      body: JSON.stringify({
+        description: formData.get("description"),
+        answer_required: formData.get("answer_required") == "on" ? true : false,
+      }),
+    },
+  });
+
+  const json = await response.json();
+
+  if (response.ok) {
+    actionFlash("success", json.message);
+  }
+
+  return ApiResponse.parse(json);
+}
+
+export async function removeQuestionAction(
+  templateId: number,
+  questionId: number,
+): Promise<ApiResponse> {
+  const response = await api({
+    input: `/manager/form-templates/${templateId}/short-questions/${questionId}`,
+    init: {
+      method: "DELETE",
+    },
+  });
+
+  const json = await response.json();
+
+  if (response.ok) {
+    actionFlash("success", json.message);
+  }
 
   return ApiResponse.parse(json);
 }
