+++ src/app/(auth)/manager/organizations/[organizationId]/form-templates/[templateId]/questions/page.tsx
@@ -1,9 +1,9 @@
 import { ActionFlashes } from "@/action-flash/ActionFlashes";
 import { VBox } from "@/components/containers";
-import { FormAddQuestion } from "@/components/pages/Questions/FormAddQuestion";
 import { Heading, Skeleton } from "@/components/ui";
 import { FormTemplate } from "@/schemas";
 import { fetchFormTemplateAction } from "../_actions";
+import { FormAddQuestion } from "./_form-add-question";
 import QuestionList from "./_questions";
 
 interface PageProps {
